backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
email-validator==2.2.0
aiosmtplib==3.0.2
jinja2==3.1.4
stripe==10.12.0
scikit-learn==1.5.2
numpy==2.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.2
edge-tts>=7.2.8
aiofiles==23.2.1
beautifulsoup4==4.12.3
lxml==5.3.0
